requirements.txt
anyio==4.0.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
pygame==2.5.2
python-telegram-bot==20.6
pytz==2023.3.post1
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
tornado==6.3.3
tzlocal==5.2
